ai-interviewer/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
dashscope==1.14.1
pdfplumber==0.10.4
python-docx==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
sse-starlette==2.0.0
python-dotenv==1.0.0